Taken From DBS Branch-Merged 45 Commits{First Commit: 61a7d2fdefc43da8c26dff306c0aac538f92a6c8} to {Last Commit: d46f04f18e43d0bdf05df50e282784d9e40030db}

+++ core/src/main/java/org/apache/carbondata/core/util/ByteUtil.java
@@ -584,9 +584,9 @@ public final class ByteUtil {
     if (CarbonUnsafe.getUnsafe() != null) {
       if (CarbonUnsafe.IS_LITTLE_ENDIAN) {
         l = Long.reverseBytes(
-            CarbonUnsafe.getUnsafe().getLong(bytes, offset + CarbonUnsafe.BYTE_ARRAY_OFFSET));
+            CarbonUnsafe.getUnsafe().getLong(bytes, (long)(offset + CarbonUnsafe.BYTE_ARRAY_OFFSET)));
       } else {
-        l = CarbonUnsafe.getUnsafe().getLong(bytes, offset + CarbonUnsafe.BYTE_ARRAY_OFFSET);
+        l = CarbonUnsafe.getUnsafe().getLong(bytes, (long)(offset + CarbonUnsafe.BYTE_ARRAY_OFFSET));
       }
     } else {
       for (int i = offset; i < offset + length; i++) {
